backend/requirements.txt
fastapi
uvicorn
pygithub
beautifulsoup4
requests
retry
python-multipart
pydantic
